Move global static mutexes into function static mutexes
ICU uses a few global static initializers for synchronizing access to
data, locale, etc. These static variables are not properly initialized
when doing link time optimizations for Windows as initialization of
static global variables is dependent on linker behavior. Move these
static globals into function static members so initialization happens
exactly once when they are first used.

Note: This fix was eventually implemented upstream as
https://github.com/unicode-org/icu/commit/bc4bb89af274f17ef28ec7bccd4c25c696120d6a.
It could not cleanly be cherry picked as the upstream fix is on a later
version of ICU with a lot of divergence. For the next ICU update, this
commit should be reverted before merging upstream master to be in sync
with ICU going forward.