Allow winuwp to read data from data in memory (#2)
The runtime uses data by converting .dat files into memory models which
are compiled directly into the ICU static libraries and linked into the final
runtimecore shared library. This is done to keep the runtime completely
portable and stand alone which is a requirement for mobile platforms.
ICU has code paths for WinUWP that avoid the ability to read data from
memory, due to the constraints that UWP adds to loading because of the
UWP sandbox. Unfortunately, limiting the loading of dlls also impacts the
ability to read from memory. Avoid these code paths for data loading so
the memory data can be utilized.

Reviewed-by: Dan Kalowsky <dkalowsky@esri.com>
Signed-off-by: Christian O. Venegas <cvenegas@esri.com>