ICU-21662 UVector::insertElementAt() error handling
Change the error handling of UVector::insertElementAt() to delete an incoming
object that would otherwise be adopted if the insert fails for any reason.
Also, change the function to never insert when there is an incoming error condition.

Adjust call sites as needed, including in transliterator files and serv.cpp

For files affected by the insertElementAt() changes, also check any uses of
the old UVector::addElementX(), changing to the new addElement().