Implement default formatters by type
Adds a new function registry method, getFormatterByType(), and a builder
method for registering formatters by type. This is analogous to ICU4J's
`getDefaultFormatterNameForType()`.

The `FormattableObject` type tags are used as keys into a map, so as part
of this change, `tag()` now returns a UnicodeString instead of an STL
string view type, since this method is now used internally.

Also made some unary constructors `explicit`.